more nodes

+++ elasticsearch/docker-entrypoint.sh
@@ -7,14 +7,38 @@ if [ "${1:0:1}" = '-' ]; then
 	set -- elasticsearch "$@"
 fi
 
-rm -rf /tmp/ssl
-mkdir -p /tmp/ssl
-keytool -genkey -keyalg RSA -noprompt -alias $SERVERNAME -dname "CN=$SERVERNAME,OU=IDM,O=EEA,L=IDM1,C=DK" -keystore /tmp/ssl/self.jks -storepass $KIBANA_RW_PASSWORD -keypass $KIBANA_RW_PASSWORD
-keytool -keystore  /tmp/ssl/self.jks -alias $SERVERNAME -export -file  /tmp/ssl/self.cert
+if [ $ELASTICSEARCH_MASTER = "YES" ]
+then
+  cat /readonlyrest.yml >> /usr/share/elasticsearch/config/elasticsearch.yml
+  bin/elasticsearch-plugin install file:/usr/share/elasticsearch/readonlyrest.zip
 
-chmod 400 /tmp/ssl/*
+  rm -rf /tmp/ssl
+  mkdir -p /tmp/ssl
+  keytool -genkey -keyalg RSA -noprompt -alias $SERVERNAME -dname "CN=$SERVERNAME,OU=IDM,O=EEA,L=IDM1,C=DK" -keystore /tmp/ssl/self.jks -storepass $KIBANA_RW_PASSWORD -keypass $KIBANA_RW_PASSWORD
+  keytool -keystore  /tmp/ssl/self.jks -alias $SERVERNAME -export -file  /tmp/ssl/self.cert
 
-sed "s#CHECKHEALTH#$CHECKHEALTH#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
+  chmod 400 /tmp/ssl/*
+
+  sed "s#CHECKHEALTH#$CHECKHEALTH#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
+
+  echo "node.master: true" >> /usr/share/elasticsearch/config/elasticsearch.yml
+
+  sed "s#KIBANA_RW_USERNAME#$KIBANA_RW_USERNAME#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
+  sed "s#KIBANA_RW_PASSWORD#$KIBANA_RW_PASSWORD#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
+
+  sed "s#LOGSTASH_RW_USERNAME#$LOGSTASH_RW_USERNAME#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
+  sed "s#LOGSTASH_RW_PASSWORD#$LOGSTASH_RW_PASSWORD#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
+
+  if [ "$KIBANA_RO_USERNAME" ]; then
+    sed "s#KIBANA_RO_USERNAME#$KIBANA_RO_USERNAME#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
+    sed "s#KIBANA_RO_PASSWORD#$KIBANA_RO_PASSWORD#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
+  fi
+ 
+else
+  echo "node.master: false" >> /usr/share/elasticsearch/config/elasticsearch.yml
+  echo 'discovery.zen.ping.unicast.hosts: ["ELASTICSEARCH_MASTER"]' >> /usr/share/elasticsearch/config/elasticsearch.yml
+  sed "s#ELASTICSEARCH_MASTER#$ELASTICSEARCH_MASTER#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
+fi
 
 # Drop root privileges if we are running elasticsearch
 # allow the container to be started with `--user`
@@ -31,17 +55,6 @@ if [ "$1" = 'elasticsearch' -a "$(id -u)" = '0' ]; then
 	#exec gosu elasticsearch "$BASH_SOURCE" "$@"
 fi
 
-sed "s#KIBANA_RW_USERNAME#$KIBANA_RW_USERNAME#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
-sed "s#KIBANA_RW_PASSWORD#$KIBANA_RW_PASSWORD#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
-
-sed "s#LOGSTASH_RW_USERNAME#$LOGSTASH_RW_USERNAME#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
-sed "s#LOGSTASH_RW_PASSWORD#$LOGSTASH_RW_PASSWORD#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
-
-if [ "$KIBANA_RO_USERNAME" ]; then
-	sed "s#KIBANA_RO_USERNAME#$KIBANA_RO_USERNAME#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
-	sed "s#KIBANA_RO_PASSWORD#$KIBANA_RO_PASSWORD#g" -i /usr/share/elasticsearch/config/elasticsearch.yml
-	sed "s#\#-RO-##g" -i /usr/share/elasticsearch/config/elasticsearch.yml
-fi
 
 # As argument is not related to elasticsearch,
 # then assume that user wants to run his own process,
